Use multiset in similarity algorithm
We recently had the answer "a hostile hostel," to which I answered just
"hostel". Since there are a lot of shared letter pairs between the two
words (somewhat intentionally, since the category was about homophones),
it gave it to me when I really shouldn't have gotten it.

By using a multiset, we make sure we don't ignore the effect of repeated
letter pairs on the answer. This would probably work worse for general
search, but better for trying to match trivia answers.