Send updates as fairly as possible
Before this commit, WebSocket messages were built and sent serially. If
some messages are sufficiently time-intensive to produce, updates could
arrive with meaningful time differences between them. This could impact
fairness.

This commit changes the logic to compute and send all the messages in
parallel, with a barrier in place to wait for all threads to have
produced their message before any of them start sending. This should be
as fair as I can possibly make it.